requirements.txt
cudatoolkit==11.3.1
cudnn==8.2.1
cycler==0.12.1
keras==2.6.0
keras-preprocessing==1.1.2
matplotlib==3.4.3
numpy==1.26.4
opencv-python==4.9.0.80
openslide==3.4.1
openslide-python==1.3.1
pandas==2.2.2
pillow==10.3.0
pip==23.3.1
python==3.9.19
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.13.1
tensorboard==2.16.2
tensorflow-gpu==2.6.0
termcolor==1.1.0
tqdm==4.66.2
vit==2.3.2
vit-keras==0.1.2